Move dialog functions to separate file
